add notion of vm state to JitBuilder
The primary purpose of JitBuilder is to help developers create Jits for
language runtimes. Jit compilers need to understand the virtual machine
state for the runtime in which the generated code will operate. In
particular, it is often necessary for the generated code to call code
that runs in the interpreter. One way to do that is for the generated
code to fully maintain the interpreter's virtual machine state at
all times. It is possible to that today with JitBuilder, but it's not
the most effective way to write a Jit compiler because maintaining
that state all the time isn't always efficient because the compiler
cannot always optimize the virtual machine state updates.

A more effective approach is for the JIT compiler to simulate the
virtual machine state as it is generating the code for each bytecode
and to only "commit" that simulated state to the interpreter's
virtual machine state when a transition has to occur.

This commit introduces an interface called VirtualMachineState that
can be used as part of this simulation process. I also created three
kinds of virtual machine state implementations: one for virtual
machine register (basically a value stored at some address in the
virtual machine), one for a virtual machine register that resides
in some readily accessible struct (for example, a "thread" structure
passed as a parameter to the compiled method), and one for an
operand stack as used in many bytecode-based virtual machines.

A Jit created using JitBuilder can subclass these structures to
create their own representation of the virtual machine state for
their language runtime, and then describe it to JitBuilder.
The BytecodeBuilder and MethodBuilder objects have also been
augmented to automatically create and propagate virtual machine
state while a method is being compiled via JitBuilder. At any
time that control will flow to the interpreter, the language Jit
can simply call the Commit() function and pass the builder object
representing the execution path that leads to the interpreter,
and the operations needed to make the interpreter's virtual
machine state match the simulated state will be automatically
appended to that builder object.

A subsequent commit will add tests for some of these new
structures.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>