Add bcLength to BytecodeBuilder
The bcLength is the number of bytes this builder consumes from the
bytecode stream. Use this to set PC in the interpreter and to add
fallthrough builders in the JIT.

Add DefaultFallThroughTarget and ReturnTarget to RuntimeBuilder to
control bytecode flow in the interpreter and JIT. I still have to
figure out what to do for Calls in my interpreter / JIT since it
is currently hacked.

Rename JumpBuilder to JumpIfBuilder

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>